+++ src/main/java/it/cavallium/dbengine/database/disk/LLLocalMultiLuceneIndex.java
@@ -203,7 +203,7 @@ public class LLLocalMultiLuceneIndex implements LLLuceneIndex {
 	@Override
 	public Mono<LLSearchResult> moreLikeThis(@Nullable LLSnapshot snapshot,
 			Flux<Tuple2<String, Set<String>>> mltDocumentFields,
-			int limit,
+			long limit,
 			@Nullable Float minCompetitiveScore,
 			String keyFieldName) {
 		long actionId;
@@ -272,7 +272,7 @@ public class LLLocalMultiLuceneIndex implements LLLuceneIndex {
 	@Override
 	public Mono<LLSearchResult> search(@Nullable LLSnapshot snapshot,
 			Query query,
-			int limit,
+			long limit,
 			@Nullable LLSort sort,
 			LLScoreMode scoreMode,
 			@Nullable Float minCompetitiveScore,
